@@ -3,10 +3,8 @@ package engine_test
 import (
 	"context"
 	"errors"
-	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -17,18 +15,9 @@ import (
 )
 
 func TestGenerate(t *testing.T) {
-	ctx := context.Background()
+	ctx := t.Context()
 
 	nooparser := func(context.Context, string, *testconfig) error { return nil }
-
-	t.Run("error_missing_parsers_templates", func(t *testing.T) {
-		// Act
-		_, err := engine.Generate(ctx, testconfig{})
-
-		// Assert
-		assert.ErrorIs(t, err, engine.ErrMissingParsers)
-		assert.ErrorIs(t, err, engine.ErrMissingTemplates)
-	})
 
 	t.Run("error_parsing", func(t *testing.T) {
 		// Arrange
@@ -36,27 +25,10 @@ func TestGenerate(t *testing.T) {
 		parser := func(context.Context, string, *testconfig) error { return errors.New("some error") }
 
 		// Act
-		_, err := engine.Generate(ctx, testconfig{},
-			engine.WithDestination[testconfig](destdir),
-			engine.WithTemplates(os.DirFS(destdir), []engine.Template[testconfig]{{}}),
-			engine.WithParsers(parser))
+		_, err := engine.Generate(ctx, destdir, testconfig{}, []engine.Parser[testconfig]{parser}, nil)
 
 		// Assert
 		assert.ErrorContains(t, err, "some error")
-	})
-
-	t.Run("error_missing_out", func(t *testing.T) {
-		// Arrange
-		destdir := t.TempDir()
-
-		// Act
-		_, err := engine.Generate(ctx, testconfig{},
-			engine.WithDestination[testconfig](destdir),
-			engine.WithTemplates(os.DirFS(destdir), []engine.Template[testconfig]{{}}),
-			engine.WithParsers(nooparser))
-
-		// Assert
-		assert.ErrorIs(t, err, engine.ErrMissingOut)
 	})
 
 	t.Run("error_read_template_out", func(t *testing.T) {
@@ -66,28 +38,12 @@ func TestGenerate(t *testing.T) {
 		require.NoError(t, os.Mkdir(filepath.Join(destdir, template.Out), files.RwxRxRxRx))
 
 		// Act
-		_, err := engine.Generate(ctx, testconfig{},
-			engine.WithDestination[testconfig](destdir),
-			engine.WithTemplates(os.DirFS(destdir), []engine.Template[testconfig]{template}),
-			engine.WithParsers(nooparser))
+		_, err := engine.Generate(ctx, destdir, testconfig{},
+			[]engine.Parser[testconfig]{nooparser},
+			[]engine.Generator[testconfig]{engine.GeneratorTemplates(os.DirFS(destdir), []engine.Template[testconfig]{template})})
 
 		// Assert
-		assert.ErrorContains(t, err, "should generate")
-	})
-
-	t.Run("error_template_invalid_globs", func(t *testing.T) {
-		// Arrange
-		destdir := t.TempDir()
-		template := engine.Template[testconfig]{Out: "file.txt"}
-
-		// Act
-		_, err := engine.Generate(ctx, testconfig{},
-			engine.WithDestination[testconfig](destdir),
-			engine.WithTemplates(os.DirFS(destdir), []engine.Template[testconfig]{template}),
-			engine.WithParsers(nooparser))
-
-		// Assert
-		assert.ErrorIs(t, err, engine.ErrMissingGlobs)
+		assert.ErrorIs(t, err, engine.ErrFailedGeneration)
 	})
 
 	t.Run("error_parse_template_globs", func(t *testing.T) {
@@ -99,34 +55,12 @@ func TestGenerate(t *testing.T) {
 		}
 
 		// Act
-		_, err := engine.Generate(ctx, testconfig{},
-			engine.WithDestination[testconfig](destdir),
-			engine.WithTemplates(os.DirFS(destdir), []engine.Template[testconfig]{template}),
-			engine.WithParsers(nooparser))
+		_, err := engine.Generate(ctx, destdir, testconfig{},
+			[]engine.Parser[testconfig]{nooparser},
+			[]engine.Generator[testconfig]{engine.GeneratorTemplates(os.DirFS(destdir), []engine.Template[testconfig]{template})})
 
 		// Assert
-		assert.ErrorContains(t, err, "parse template file(s)")
-	})
-
-	t.Run("success_template_already_exists", func(t *testing.T) {
-		// Arrange
-		destdir := t.TempDir()
-		template := engine.Template[testconfig]{Out: "file.txt"}
-		require.NoError(t, os.WriteFile(filepath.Join(destdir, template.Out), []byte("some not empty file"), files.RwRR))
-
-		buf := strings.Builder{}
-		logger := engine.NewTestLogger(&buf)
-
-		// Act
-		_, err := engine.Generate(ctx, testconfig{},
-			engine.WithLogger[testconfig](logger),
-			engine.WithDestination[testconfig](destdir),
-			engine.WithTemplates(os.DirFS(destdir), []engine.Template[testconfig]{template}),
-			engine.WithParsers(nooparser))
-
-		// Assert
-		require.NoError(t, err)
-		assert.Equal(t, buf.String(), fmt.Sprintf("not generating '%s' since it already exists", template.Out))
+		assert.ErrorIs(t, err, engine.ErrFailedGeneration)
 	})
 
 	t.Run("success_template", func(t *testing.T) {
@@ -146,15 +80,14 @@ func TestGenerate(t *testing.T) {
 		require.NoError(t, outfile.Close())
 
 		// Act
-		_, err = engine.Generate(ctx, testconfig{},
-			engine.WithDestination[testconfig](destdir),
-			engine.WithTemplates(os.DirFS(destdir), []engine.Template[testconfig]{template}),
-			engine.WithParsers(nooparser))
+		_, err = engine.Generate(ctx, destdir, testconfig{},
+			[]engine.Parser[testconfig]{nooparser},
+			[]engine.Generator[testconfig]{engine.GeneratorTemplates(os.DirFS(destdir), []engine.Template[testconfig]{template})})
 
 		// Assert
 		require.NoError(t, err)
-		bytes, err := os.ReadFile(out)
+		content, err := os.ReadFile(out)
 		require.NoError(t, err)
-		assert.Equal(t, "value  is empty, since no parser updated it", string(bytes))
+		assert.Equal(t, "value  is empty, since no parser updated it", string(content))
 	})
 }
