@@ -10,12 +10,11 @@ import (
 
 // Dependabot returns the slice of templates related to dependabot configuration.
 func Dependabot() []engine.Template[craft.Config] {
-	name := path.Join(".github", "dependabot.yml")
 	return []engine.Template[craft.Config]{
 		{
 			Delimiters: engine.DelimitersBracket(),
-			Globs:      []string{name + engine.TmplExtension},
-			Out:        name,
+			Globs:      []string{path.Join(".github", "dependabot.yml") + engine.TmplExtension},
+			Out:        path.Join(".github", "dependabot.yml"),
 			Remove: func(config craft.Config) bool {
 				return config.Platform != parser.GitHub || !config.IsBot(craft.Dependabot)
 			},
@@ -25,25 +24,20 @@ func Dependabot() []engine.Template[craft.Config] {
 
 // Renovate returns the slice of templates related to renovate configuration.
 func Renovate() []engine.Template[craft.Config] {
-	var templates []engine.Template[craft.Config]
-
-	yml := path.Join(".github", "workflows", "renovate.yml")
-	templates = append(templates, engine.Template[craft.Config]{
-		Delimiters: engine.DelimitersChevron(),
-		Globs:      []string{yml + engine.TmplExtension},
-		Out:        yml,
-		Remove: func(config craft.Config) bool {
-			return !config.IsBot(craft.Renovate) || !config.IsCI(parser.GitHub) || (config.CI.Auth.Maintenance != nil && *config.CI.Auth.Maintenance == craft.Mendio) //nolint:revive
+	return []engine.Template[craft.Config]{
+		{
+			Delimiters: engine.DelimitersChevron(),
+			Globs:      []string{path.Join(".github", "workflows", "renovate.yml") + engine.TmplExtension},
+			Out:        path.Join(".github", "workflows", "renovate.yml"),
+			Remove: func(config craft.Config) bool {
+				return !config.IsBot(craft.Renovate) || !config.IsCI(parser.GitHub) || (config.CI.Auth.Maintenance != nil && *config.CI.Auth.Maintenance == craft.Mendio) //nolint:revive
+			},
 		},
-	})
-
-	json5 := "renovate.json5"
-	templates = append(templates, engine.Template[craft.Config]{
-		Delimiters: engine.DelimitersChevron(),
-		Globs:      []string{json5 + engine.TmplExtension},
-		Out:        json5,
-		Remove:     func(config craft.Config) bool { return !config.IsBot(craft.Renovate) },
-	})
-
-	return templates
+		{
+			Delimiters: engine.DelimitersChevron(),
+			Globs:      []string{"renovate.json5" + engine.TmplExtension},
+			Out:        "renovate.json5",
+			Remove:     func(config craft.Config) bool { return !config.IsBot(craft.Renovate) },
+		},
+	}
 }
