@@ -31,12 +31,12 @@ var generated = regexp.MustCompile(`Code generated by [\w\-\/]+; DO NOT EDIT.`)
 //   - it does not exist
 //   - it is empty
 //   - the policy is set to PolicyAlways
-func ShouldGenerate(dst string, policy Policy) (bool, error) {
+func ShouldGenerate(out string, policy Policy) (bool, error) {
 	if policy == PolicyAlways {
 		return true, nil
 	}
 
-	content, err := os.ReadFile(dst)
+	content, err := os.ReadFile(out)
 	if err != nil {
 		if errors.Is(err, fs.ErrNotExist) {
 			return true, nil
