@@ -5,24 +5,39 @@ Examples:
 
 	type config struct { ... }
 
-	func Parser(ctx context.Context, destdir string, c *config) error {
-		// do something to parse the repository / configuration
+	func ParserNode(ctx context.Context, destdir string, c *config) error {
+		var jsonfile parser.PackageJSON
+		jsonpath := filepath.Join(destdir, parser.FilePackageJSON)
+		if err := files.ReadJSON(jsonpath, &jsonfile, os.ReadFile); err != nil {
+			if errors.Is(err, fs.ErrNotExist) {
+				return nil
+			}
+			return fmt.Errorf("read json: %w", err)
+		}
+		engine.GetLogger().Infof("node detected, a '%s' is present and valid", parser.FilePackageJSON)
+
+		if err := jsonfile.Validate(); err != nil {
+			return fmt.Errorf("validate '%s': %w", parser.FilePackageJSON, err)
+		}
+		// do something with parsed jsonfile (e.g. update config since it's a pointer)
 		return nil
 	}
 
-	var _ generate.Parser[config] = Parser // ensure interface is implemented
+	var _ generate.Parser[config] = ParserNode // ensure interface is implemented
 
 	// single parser call
 	func main() {
 		var c config
-		err := parser.Parser(ctx, "path/to/dir", &c)
+		err := parser.ParserNode(ctx, "path/to/dir", &c)
 		// handle err
 	}
 
 	// fully used with engine.Generate
 	func main() {
+		destdir, _ := os.Getwd()
+
 		var c config
-		err := engine.Generate(ctx, &c, generate.WithParsers(Parser, ...))
+		err := engine.Generate(ctx, destdir, c, []engine.Parser[config]{ParserNode})
 		// handle err
 	}
 */
